+++ src/main/java/mx/com/lestradam/algorithms/functions/fitness/FFGeneticAlgorithm.java
@@ -11,7 +11,7 @@ public class FFGeneticAlgorithm {
 	@Autowired
 	private ObjectiveFunction objFunction;
 
-	public long evaluateSolution(long[] solution){
+	public double evaluateSolution(long[] solution){
 		return objFunction.evaluate(solution);
 	}
