+++ src/main/java/mx/com/lestradam/algorithms/functions/fitness/ObjectiveFunction.java
@@ -2,6 +2,6 @@ package mx.com.lestradam.algorithms.functions.fitness;
 
 public interface ObjectiveFunction {
 	
-	public long evaluate(long[] solution);
+	public double evaluate(long[] solution);
 
 }
