+++ src/test/java/mx/com/lestradam/algorithms/functions/basic/StatisticalOperationsTest.java
@@ -4,13 +4,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import mx.com.lestradam.algorithms.functions.basic.StatisticalOperations;
-
 class StatisticalOperationsTest {
 
 	@Test
 	void testGetAvgValue() {
-		long[] values = new long[] {100,100,300, 102};
+		double[] values = new double[] { 100, 100, 300, 102 };
 		double avg = StatisticalOperations.getAvgValue(values);
 		assertEquals(150.5, avg);
 	}
